+++ Models/Pasivos.cs
@@ -7,7 +7,7 @@ namespace Cooperativa.Models
         [Key]
         public int pasivoID { get; set; }
         public String? tipo { get; set; }
-        public double costoPasivo { get; set; }
+        public decimal costoPasivo { get; set; }
         public String? detalle { get; set; }
         public DateTime fechaPasivo { get; set; }
     }
